refactor : 과정과 강의를 N:M으로 매핑해주는 테이블 추가
