+++ src/main/java/nextstep/courses/infrastructure/CourseSessionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.infrastructure;
 
 import java.util.List;
